# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stellar MarketPay API",
    "version": "1.0.0",
    "description": "Backend API for Stellar MarketPay - A decentralized freelance marketplace built on the Stellar blockchain.\n\n## Authentication\nMost endpoints require JWT authentication obtained via the **/api/auth** challenge-response flow using Stellar wallet signatures.\n\n## Pagination\nList endpoints support cursor-based pagination via the `after` query parameter. Responses include `next_cursor` and `has_more` fields.",
    "contact": {
      "name": "Stellar MarketPay Team",
      "email": "support@stellarmarketpay.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    },
    {
      "url": "https://api.stellarmarketpay.com",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth (Stellar wallet signature)"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "JWT cookie set after authentication"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Developer API key for public API endpoints"
      }
    },
...
  },
  "paths": {
...
    "/api/messages/job/{jobId}": {
      "post": {
        "summary": "Send a message in a job thread",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "Message content (encrypted)"
                  },
                  "contractTxHash": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent"
          }
        }
      },
      "get": {
        "summary": "Get messages for a job thread",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message list (marks as read)"
          }
        }
      }
    },
    "/api/messages/unread-count": {
      "get": {
        "summary": "Get total unread message count",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Unread count"
          }
        }
      }
    },
    "/api/messages/{messageId}/tx-hash": {
      "patch": {
        "summary": "Attach on-chain tx hash to a message",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "messageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "txHash"
                ],
                "properties": {
                  "txHash": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tx hash attached"
          }
        }
      }
    },
    "/api/messages/job/{jobId}/attachments": {
      "post": {
        "summary": "Upload an encrypted file attachment",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "senderNaclPub": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attachment uploaded to IPFS"
          }
        }
      }
    },
...
  }
}